split responsibility between navigation and login guard
