+++ src/App.tsx
@@ -1,13 +1,14 @@
 import './App.css';
-import { Routes, Route } from 'react-router-dom';
+import { Routes, Route, Navigate } from 'react-router-dom';
 import { ThemeProvider } from '@iexec/react-ui-kit';
 import {
-  Navigate,
+  Navigation,
   NewProtectedData,
   ProtectedData,
   Consent,
   EmailDapp,
   SendEmail,
+  LoginGuard,
 } from './features';
 import { PROTECTED_DATA, CONSENT, SEND_MAIL, CREATE } from './config/path';
 
@@ -15,24 +16,50 @@ function App() {
   return (
     <div className="App">
       <ThemeProvider>
+        <Navigation />
         <Routes>
-          <Route path="/" element={<Navigate />}>
-            <Route path={`/${PROTECTED_DATA}`} element={<ProtectedData />} />
-            <Route
-              path={`/${PROTECTED_DATA}/${CONSENT}/:ProtectedDataId`}
-              element={<Consent />}
-            />
-            <Route
-              path={`/${PROTECTED_DATA}/${CREATE}`}
-              element={<NewProtectedData />}
-            />
-            <Route path={`/${SEND_MAIL}`} element={<EmailDapp />} />
-            <Route
-              path={`/${SEND_MAIL}/:receiverAddress/:protectedDataAddress`}
-              element={<SendEmail />}
-            />
-          </Route>
-          <Route path="*" element={<Navigate />} />
+          <Route
+            path={`/${PROTECTED_DATA}`}
+            element={
+              <LoginGuard>
+                <ProtectedData />
+              </LoginGuard>
+            }
+          />
+          <Route
+            path={`/${PROTECTED_DATA}/${CONSENT}/:ProtectedDataId`}
+            element={
+              <LoginGuard>
+                <Consent />
+              </LoginGuard>
+            }
+          />
+          <Route
+            path={`/${PROTECTED_DATA}/${CREATE}`}
+            element={
+              <LoginGuard>
+                <NewProtectedData />
+              </LoginGuard>
+            }
+          />
+          <Route
+            path={`/${SEND_MAIL}`}
+            element={
+              <LoginGuard>
+                <EmailDapp />
+              </LoginGuard>
+            }
+          />
+          <Route
+            path={`/${SEND_MAIL}/:receiverAddress/:protectedDataAddress`}
+            element={
+              <LoginGuard>
+                <SendEmail />
+              </LoginGuard>
+            }
+          />
+          {/* default redirect */}
+          <Route path="*" element={<Navigate to={`/${PROTECTED_DATA}`} />} />
         </Routes>
       </ThemeProvider>
     </div>
